Test User bootstrapping
When testing: Users are unusual in that they know things about
themselves that the application doesn't (and cannot) know. In particular
their password. So here we create a TUser class to encapsulate that
concept, and give it some default funtctionality so that each TUser has
things like: name, email, password etc.

The method missing 'cleverness' allows you to add other attributes to a
user e.g. age, without having to modify the helper class.

In our test world we can establish certain conventions about users which
help make testing easier. For example saying that users have unique
first names makes it much easier to write scenarios, as we can just
refer to the user by their first name. The TUser class will support
these conventions.

Another convention we are establishing here is the use if the @i
variable in our test world. @i will always refer to a person. That
person will be the subject of the scenario. Later when we have scenarios
that involve interactions between people we will extend this convention
to create @frank and @jill etc.