Implement sign up
Sign up is a pre-requisite for sign in.

This commit has quite alot of stuff to do with implementing sign up/in.
We use devize to get alot of functionality quickly, so we can focus more
on how to use Cukes to drive the development without being swamped by
the details of the implementation

- add devise gem
- follow instructions from https://github.com/plataformatec/devise. This
  requires care.
  - add User model
  - add migrations, routes etc.

- specifically add selenium-webdriver gem so we can work with upto date
  firefox. This allows us to use the @javascript tag to see our work in
  progress in action.