Sign in, preperation
- write happy path feature
- use @wip and @javascript tags so we can 'see' the scenario in action
- implement step defs
- delegate 'how' we sign in to helper methods

We now are at a stage where we can think about how we want the user to
sign in. This will change as the application develops. For example alot
of sites want to make you sign in as soon as you hit the site and remain
signed in between visit; whilst many others prefer to allow you to do
lots of stuff before you have to 'commit' to signing in. With the
structure we now have, such changes to 'how' we sign in will only effect
a very small number of places (basically the method bodies of our
helpers).