We can now sign in
Here we have introduced our first helper method that bypasses the UI.
This method creates a registered user. Previously we created code that
allows us to sign up through the UI. We could do this every time we
need a user, but this would be

- expensive in run time
- problematic because it would create a session

Instead we bypass the UI and add our registered user directly into the
database. Such methods should only be used in Givens.

Other things in this commit include

- fix bug in welcome page
- step defs for signing in
- removed first/last name from TUser, we don't need this yet